+++ openlane/state/__init__.py
@@ -19,5 +19,5 @@ This module manages the State of a Design before and after the execution of an
 OpenLane step. The State is essentially a list of views in various formats in
 addition to the cumulative set of metrics created by a given step.
 """
-from .design_format import DesignFormat
+from .design_format import DesignFormat, DesignFormatByID
 from .state import State